some logging added to the installer; logging functionality is extended an bits are re-factored; changes to the NewLineSequence - now there are both Unicode and mbs versions, writnl functions use sequence matching the argument type and use default build one for void* arguments; changes to installer to distinguish between an admin and non-admin users: the functionality, location of shortcuts and default installation folders differ;

+++ xlib/symmcon.cpp
@@ -133,7 +133,7 @@ SymmCon SortedSymmConTab[64] = {
 };
 //..............................................................................
 SiteSymmCon::SiteSymmCon() : added(false)  {
-  for( size_t i=0; i < 9; i++ )  {
+  for( int i=0; i < 9; i++ )  {
     map[i].param = (i < 6 ? i : i-6);
     map[i].multiplier = 1.0;
   }
